[process] Move kthread_exit() to the portable code in process.c
